Move RMC + generic functions out of misbad/MPC utils
